requirements.txt
fastapi==0.96.1
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.53.0
google-cloud-bigquery==3.21.0
google-cloud-bigquery-storage==2.24.0
google-cloud-storage==2.16.0
langchain
